requirements.txt
certifi
chardet
cycler
dataclasses
easydict~=1.11
idna
imutil
kiwisolver
matplotlib~=3.7.2
numpy~=1.26.0
opencv-python~=4.8.1.78
Pillow~=10.0.1
pkg-resources
pyparsing
python-dateutil
pyyaml~=6.0.1
requests~=2.31.0
scipy~=1.11.3
six
tqdm~=4.65.0
typing-extensions
urllib3
pandas~=2.1.1
seaborn~=0.12.2
pyqt5~=5.15.9
yaml~=0.2.5
torchvision~=0.15.2
joblib~=1.2.0
mediapipe